Move SerialBuffer class back into Esp8266 driver Component
